refactor tooltip component
