+++ src/components/tooltip/index.tsx
@@ -1,1 +1,1 @@
-export * from "./Tooltip";
+export { Tooltip } from "./Tooltip";
